+++ app/Http/Controllers/Admin/SliderController.php
@@ -20,7 +20,7 @@ class SliderController extends Controller
     public function index()
     {
         try {
-            $sliders = Slider::all();            
+            $sliders = Slider::all();
             return view('backend.sliders.index', compact('sliders'));
         } catch (\Exception $e) {
             return $e->getMessage();
@@ -101,11 +101,15 @@ class SliderController extends Controller
         try {
             $slider->update($request->except('image', 'banner'));
             if (isset($request['image'])) {
-                $slider->getFirstMedia('slider-image')->delete();
+                if ($slider->getFirstMedia('slider-image')) {
+                    $slider->getFirstMedia('slider-image')->delete();
+                }
                 $slider->addMediaFromRequest('image')->withResponsiveImages()->toMediaCollection('slider-image');
             }
             if (isset($request['banner'])) {
-                $slider->getFirstMedia('slider-banner')->delete();
+                if ($slider->getFirstMedia('slider-banner')) {
+                    $slider->getFirstMedia('slider-banner')->delete();
+                }
                 $slider->addMediaFromRequest('banner')->withResponsiveImages()->toMediaCollection('slider-banner');
             }
             return redirect()->route('sliders.index')->with('success', 'Slider updated successfully');
